Move api/admin under api
